Riqueza de espécies de sapos em fragmentos

Composição de espécies de anfíbios e variáveis ambientais coletados em nove localidades do noroeste paulista. Neste conjunto de dados as variáveis preditoras já estão padronizadas com média 0 e desvio padrão 1. As variáveis com "2" indicam variáveis quadráticas. As siglas representam a primeira letra do gênero e as duas primeiras letras do epíteto da espécie.

Composição de espécies de anfíbios e variáveis ambientais coletados em nove localidades do noroeste paulista. Neste conjunto de dados as variáveis preditoras já estão padronizadas com média 0 e desvio padrão 1. As variáveis com "2" indicam variáveis quadráticas. As siglas representam a primeira letra do gênero e as duas primeiras letras do epíteto da espécie.

fragmentos

fragmentos

Format

Data frame com 40 linhas e 40 colunas:

Riqueza_obs

Riqueza observada

Riqueza_HB

Riqueza estimada

X

Longitude

Y

Latitude

dfrag

Distância para o fragmento mais próximo

locality

Nome do município

site

Código do fragmento de mata

hydrop

Hidroperíodo do corpo d'água amostrado

vegcov

Cobertura de vegetação aquática

nveg

Número de tipos de vegetação aquática

area

Área do corpo d'água

depth

Profundidade do corpo d'água

forcov

Cobertura de floresta nativa ao redor do corpo d'água

dwater

Distância pra o corpo d'água mais próximo

Data frame com 40 linhas e 40 colunas:

Riqueza_obs

Riqueza observada

Riqueza_HB

Riqueza estimada

X

Longitude

Y

Latitude

dfrag

Distância para o fragmento mais próximo

locality

Nome do município

site

Código do fragmento de mata

hydrop

Hidroperíodo do corpo d'água amostrado

vegcov

Cobertura de vegetação aquática

nveg

Número de tipos de vegetação aquática

area

Área do corpo d'água

depth

Profundidade do corpo d'água

forcov

Cobertura de floresta nativa ao redor do corpo d'água

dwater

Distância pra o corpo d'água mais próximo

Source

https://link.springer.com/article/10.1007/s00267-014-0241-y

https://link.springer.com/article/10.1007/s00267-014-0241-y

Details

fragmentos

References

Prado, V.H.M., Rossa-Feres, D.C. (2014) Multiple Determinants of Anuran Richness and #'Occurrence #'in an Agricultural Region in South-Eastern Brazil. Environmental Management 53, #'823–837. https://doi.org/10.1007/s00267-014-0241-y

Prado, V.H.M., Rossa-Feres, D.C. (2014) Multiple Determinants of Anuran Richness and #'Occurrence #'in an Agricultural Region in South-Eastern Brazil. Environmental Management 53, #'823–837. https://doi.org/10.1007/s00267-014-0241-y