Autotomia da cauda de lagarto

Dados sobre comprimento do corpo, cauda, estado da cauda, e sexo de indivíduos do lagarto Coleodactylus meridionalis.

lagartos

Format

Data frame com 223 linhas e 6 variáveis

Numero

Número do indivíduo

Sex

Sexo do lagarto

SVL

Snout-vent length, Comprimento rostro-cloacal, uma medida de tamanho de corpo

Intact_tail_length

Comprimento da cauda intacta

Autotomized_tail_length

Comprimeiro da cauda que foi autotomizada

Tail_state

Estado da cauda, 0 = Autotomizada; 1 = cauda intacta

Source

https://www.tandfonline.com/doi/abs/10.1080/00222933.2020.1840641

References

Camila N. Oliveira, Ikaro H. M. P. Campos, Diogo B. Provete, Míriam C. Guarnieri & Samuel #'C. Ribeiro (2020)Defensive behaviour and tail autotomy in Coleodactylus meridionalis #'(Squamata: Sphaerodactylidae), Journal of Natural History, 54:33-34, 2209-2218, DOI: #'10.1080/00222933.2020.1840641