Regressões

Temperatura média anual, precipitação anual e o tamanho médio do comprimento rostro-cloacal (CRC) de populações de Dendropsophus minutus (Anura:Hylidae) amostradas em 109 localidades no Brasil.

Temperatura média anual, precipitação anual e o tamanho médio do comprimento rostro-cloacal (CRC) de populações de Dendropsophus minutus (Anura:Hylidae) amostradas em 109 localidades no Brasil.

regressoes

regressoes

Format

Um data.frame com 109 linhas e quatro colunas

Municipio

Nome dos municípios

CRC(milímetros

média do comprimento rostro-cloacal dos indivíduos em cada município

Temperatura(°C)

temperatura média dos municípios

Precipitação(milímetros)

precipitação anual nos municípios

Um data.frame com 109 linhas e quatro colunas

Municipio

Nome dos municípios

CRC(milímetros

média do comprimento rostro-cloacal dos indivíduos em cada município

Temperatura(°C)

temperatura média dos municípios

Precipitação(milímetros)

precipitação anual nos municípios

Details

regressoes

References

Boaratti, A. Z. & da Silva, F. R. (2015). Relationships between environmental gradients and geographic variation in the intraspecific body size of three species of frogs (Anura).” Austral Ecology, 40 (8), 869–76. https://doi.org/10.1111/aec.1226

Boaratti, A. Z. & da Silva, F. R. (2015). Relationships between environmental gradients and geographic variation in the intraspecific body size of three species of frogs (Anura).” Austral Ecology, 40 (8), 869–76. https://doi.org/10.1111/aec.1226