Teste t com variâncias diferentes

Comprimento rostro-cloacal (CRC) de fêmeas de Leptodactylus podicipinus amostradas em diferentes estações do ano com armadilhas de interceptação e queda na Região Noroeste do estado de São Paulo

Comprimento rostro-cloacal (CRC) de fêmeas de Leptodactylus podicipinus amostradas em diferentes estações do ano com armadilhas de interceptação e queda na Região Noroeste do estado de São Paulo

teste_t_var_diferente

teste_t_var_diferente

Format

Um data.frame com 12 linhas e duas colunas

CRC (milímetros)

Comprimento rostro-cloacal de cada indivíduo de L. podicipinus

Estacao

Tratamentos considerando as estações do ano (Chuvosa ou Seca)

Um data.frame com 12 linhas e duas colunas

CRC (milímetros)

Comprimento rostro-cloacal de cada indivíduo de L. podicipinus

Estacao

Tratamentos considerando as estações do ano (Chuvosa ou Seca)

Details

teste_t_var_diferente

References

da Silva, F. R., & Rossa-Feres, D. C. ( 2010). “Seasonal variation in body size of tropical anuran amphibians.” Herpetology Notes, 3, 205–9.

da Silva, F. R., & Rossa-Feres, D. C. ( 2010). “Seasonal variation in body size of tropical anuran amphibians.” Herpetology Notes, 3, 205–9.