Células sanguíneas de girinos

Dados de um experimento que testou como diferentes tempos de exposição à luz ultra-violeta altera as células sanguíneas em girinos de LIthobates catesbeianus com e sem pigmentação.

uv_cells

Format

Data frame com 50 linhas e 8 variáveis

UV

Tratamentos, tempos de exposição à luz Ultravioleta: Controle (CT), 6h, 12h, 18h, 24h

Pigmentation

Fator com dois níveis: codifica se o girino tinha ou não pigmentação

Total_Cell

Número total de células contadas

Lymphocyte

Número de linfócitos encontrados no sangue

Neutrophil

Número de neutrófilos encontrados no sangue

Basophil

Número de basófilos encontrados no sangue

Monocyte

Número de monócitos encontrados no sangue

Eosinophil

Número de eosinófilos encontrados no sangue

Source

https://www.sciencedirect.com/science/article/pii/S1011134417311727

References

Lilian Franco-Belussi, Lara Zácari Fanali, Classius De Oliveira, 2018. UV-B affects the immune system and promotes nuclear abnormalities in pigmented and non-pigmented bullfrog tadpoles. Journal of Photochemistry and Photobiology B: Biology, 180, 109-117